+++ test/test_interval.cc
@@ -35,7 +35,7 @@
 #include "test.h"
 
 using namespace Ariadne;
-using namespace Ariadne::Numeric;
+using namespace Ariadne;
 using namespace std;
 
 template class Interval<Rational>;
