+++ test/test_rational.cc
@@ -31,7 +31,7 @@
 #include "test.h"
 
 using namespace std;
-using namespace Ariadne::Numeric;
+using namespace Ariadne;
 
 class TestRational
 {
