+++ test/test_rounding.cc
@@ -32,7 +32,7 @@
 
 using namespace std;
 using namespace Ariadne;
-using namespace Ariadne::Numeric;
+using namespace Ariadne;
 
 class TestRounding 
 {
